# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoneyMatch API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "me"
        ],
        "summary": "Update Me",
        "operationId": "update_me_api_v1_me_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/self-exclude": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Post Self Exclude",
        "operationId": "post_self_exclude_api_v1_me_self_exclude_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/links": {
      "get": {
        "tags": [
          "links"
        ],
        "summary": "List Links",
        "operationId": "list_links_api_v1_links_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "links"
        ],
        "summary": "Create Link",
        "operationId": "create_link_api_v1_links_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/links/{game}/profile": {
      "get": {
        "tags": [
          "links"
        ],
        "summary": "Refresh Link",
        "operationId": "refresh_link_api_v1_links__game__profile_get",
        "parameters": [
          {
            "name": "game",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/links/{game}": {
      "delete": {
        "tags": [
          "links"
        ],
        "summary": "Delete Link",
        "operationId": "delete_link_api_v1_links__game__delete",
        "parameters": [
          {
            "name": "game",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateLinkRequest": {
        "properties": {
          "game": {
            "type": "string",
            "title": "Game"
          },
          "username": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "game",
          "username"
        ],
        "title": "CreateLinkRequest",
        "description": "Intent to bind a host account. The server fetches + verifies everything;\nthe client only names the game and the handle to claim."
      },
...
      "FormatStat": {
        "properties": {
          "speed": {
            "type": "string",
            "enum": [
              "bullet",
              "blitz",
              "rapid",
              "classical"
            ],
            "title": "Speed"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "games": {
            "type": "integer",
            "title": "Games"
          },
          "provisional": {
            "type": "boolean",
            "title": "Provisional",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "speed",
          "rating",
          "games"
        ],
        "title": "FormatStat",
        "description": "A single chess time-control's verified stats, sourced from the host."
      },
...
      "GameLink": {
        "properties": {
          "game": {
            "type": "string",
            "title": "Game"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "LINKED",
              "BLOCKED",
              "UNLINKED"
            ],
            "title": "Status"
          },
          "host_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host Username"
          },
          "linked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked At"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileSnapshot"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "game",
          "display_name",
          "status"
        ],
        "title": "GameLink",
        "description": "One game's row on Profile: linked snapshot or an empty/blocked slot.\n\n`status`: LINKED (active binding), BLOCKED (game flag off or binding frozen),\nUNLINKED (available to link)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LimitsResponse": {
        "properties": {
          "daily_loss_cap_cents": {
            "type": "integer",
            "title": "Daily Loss Cap Cents"
          },
          "daily_entry_cap_cents": {
            "type": "integer",
            "title": "Daily Entry Cap Cents"
          },
          "max_concurrent_contests": {
            "type": "integer",
            "title": "Max Concurrent Contests"
          },
          "pending_limits": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Limits"
          },
          "pending_effective_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Effective At"
          }
        },
        "type": "object",
        "required": [
          "daily_loss_cap_cents",
          "daily_entry_cap_cents",
          "max_concurrent_contests",
          "pending_limits",
          "pending_effective_at"
        ],
        "title": "LimitsResponse"
      },
      "LinksResponse": {
        "properties": {
          "games": {
            "items": {
              "$ref": "#/components/schemas/GameLink"
            },
            "type": "array",
            "title": "Games"
          }
        },
        "type": "object",
        "required": [
          "games"
        ],
        "title": "LinksResponse",
        "description": "Every registered game with the viewer's link state \u2014 drives Profile."
      },
...
      "MeResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "needs_onboarding": {
            "type": "boolean",
            "title": "Needs Onboarding"
          },
          "limits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LimitsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "unread_notifications": {
            "type": "integer",
            "title": "Unread Notifications",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "user",
          "needs_onboarding"
        ],
        "title": "MeResponse",
        "description": "`/me` payload: the user, a computed onboarding flag, and staking limits."
      },
...
      "ProfileSnapshot": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "link_method": {
            "type": "string",
            "enum": [
              "username",
              "oauth"
            ],
            "title": "Link Method"
          },
          "game": {
            "type": "string",
            "title": "Game",
            "default": "chess.lichess"
          },
          "account_age_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Age Days"
          },
          "win_rate": {
            "type": "number",
            "title": "Win Rate"
          },
          "draw_rate": {
            "type": "number",
            "title": "Draw Rate",
            "default": 0.0
          },
          "total_games": {
            "type": "integer",
            "title": "Total Games"
          },
          "formats": {
            "items": {
              "$ref": "#/components/schemas/FormatStat"
            },
            "type": "array",
            "title": "Formats",
            "default": []
          },
          "primary_speed": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "bullet",
                  "blitz",
                  "rapid",
                  "classical"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Speed"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "rank_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Label"
          },
          "kd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kd"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "username",
          "display_name",
          "url",
          "link_method",
          "win_rate",
          "total_games"
        ],
        "title": "ProfileSnapshot",
        "description": "Verified skill profile for one linked host account.\n\nChess populates the per-format / `primary_speed` fields; other titles leave\nthose empty and use the generic `rating` / `rank_label` / `kd` descriptors.\n`game` is the adapter id."
      },
...
      "UpdateMeRequest": {
        "properties": {
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "3\u201320 chars [a-z0-9_]; once"
          },
          "residence_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Residence State",
            "description": "2-letter US state"
          },
          "dob_attested_18plus": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dob Attested 18Plus"
          },
          "daily_loss_cap_cents": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Loss Cap Cents"
          },
          "daily_entry_cap_cents": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Entry Cap Cents"
          }
        },
        "type": "object",
        "title": "UpdateMeRequest",
        "description": "Onboarding (username + state + 18+, set once) and/or limit edits.\n\nAll fields optional so the endpoint serves both onboarding and later limit\nchanges. Lowering a cap is instant; raising is delayed (see limits_service)."
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "friend_code": {
            "type": "string",
            "title": "Friend Code"
          },
          "residence_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Residence State"
          },
          "dob_attested_18plus": {
            "type": "boolean",
            "title": "Dob Attested 18Plus"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "member_since": {
            "type": "string",
            "format": "date-time",
            "title": "Member Since"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "friend_code",
          "residence_state",
          "dob_attested_18plus",
          "role",
          "status",
          "member_since"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}